Extract Chat and Kmail domains from Client
Move chat and kmail methods out of the monolithic Client class into
dedicated domain classes following the existing Players pattern.

- Chat domain: check, macro, send (whisper), getUpdates
- Kmail domain: fetch, delete, check, send, static parse
- Inline utils/kmail.ts types and parsing into the domains
- Rename types: BaseKoLMessage → Message, KoLChatMessage → RawChatMessage,
  KoLKmail → RawKmail. Drop the KoLMessage union in favour of precise
  ChatMessage/KmailMessage event types.
- Update all consumers in packages/oaf
